requirements.txt
requests==2.31.0
python-decouple==3.8
pydantic==2.5.3
python-slugify==8.0.1
supabase==2.7.4
postgrest==0.16.8
schedule==1.2.0
pytz==2023.3
python-dateutil==2.8.2